Commit 7/8/2025 10:10 pm
- added check functions for king pieces, uses move_line function and
checks for every possible check spot
- added turn tracking, so black and white move alternatively.
- added king move logic, king cannot move where there is a check. all
use the check_if_valid function which houses the logic if an enemy piece
is covering that square, whether a king move or the king itself
- added logic within check_if_valid to return the moveset that the
player can either block or capture the checking piece. accounts for
double checks
- connected everything to the chess_game files so if player gives a
check, then next players only moves has to  be a move to get out of
check.

----------------
to dos
- need to implmenet a check for checkmate